Disable exports by default
Add INIT_EXPORT init flag. When set (with the KOS_INIT_FLAGS macro), the
KOS symbols will be exported. The default is to not export them.

This helps reduce the size of the executables, as the vast majority of
them don't need the symbol export feature, by telling the compiler that
it's okay to garbage-collect them.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>