Garbage-collect the network stack when unused
Make it so that the network stack (and BBA and LAN adapter code)
is garbage-collected when a KOS program has been compiled without the
INIT_NET flag (and doesn't use any of the symbols).

This really drops the size of the compiled binaries. For instance, the
dreamcast "hello" example dropped from 284 KiB to 156 KiB (both
stripped).

Signed-off-by: Paul Cercueil <paul@crapouillou.net>