Add and use new macros for init/shutdown functions
The KOS_INIT_FLAG_EXPORT() is supposed to be used within
KOS_INIT_FLAGS(). It will override the weak symbol associated with the
function with a strong one, that points to this function if the flag is
set, or is NULL if the flag is cleared.

The KOS_INIT_FLAG_WEAK() should be used to declare the weak symbol for
the init/shutdown function.

Finally, the init/shutdown function can be called with
KOS_INIT_FLAG_CALL(). If the flag has not been set, the function will
not be called, and will be automatically garbage-collected by the
compiler.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>