aicaos: Add interrupt routines
Add the basic irq_disable() and irq_restore() functions, which permit to
create critical sections.

Add the irq_disable_scoped() macro as a scoped variant as well,
irq_enabled() as a way to know if IRQs are enabled, and a function
aica_interrupt(), which can be used to send an interrupt to the SH4.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>