aicaos: tasks: Add support for sleeping
When rescheduling either manually or after an IRQ, run the next task for
a maximum of 1764 timer ticks (roughly 40 ms) before the next
reschedule, and a minimum defined by the earliest time that a
higher-priority task in sleep state is supposed to wake up.

At each reschedule, all sleeping tasks' timer is decreased by the time
since the last reschedule. If this timer drops to zero, the task's state
is set back to "running".

Signed-off-by: Paul Cercueil <paul@crapouillou.net>