aicaos: Add function to register a channel as counter
Since the AICA allows reading the position at which a given sample is
currently being played on a channel, by using a bogus sample (address
0x0 in RAM, which is just non-sample data) while muting the channel, we
have a 16-bit counter that runs at 44100 Hz.

As the AICA timers are set-and-forget, and cannot be read, (ab)using a
channel as a counter will allow for accurate time accounting.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>