maple: Fix race between IRQ and threads
The IRQ handler and threads would read-modify-write the state->queue_len
field as if races didn't exist.

Address this issue by using atomics instead.

Note that the code in the IRQ handler is using atomic routines as well
just to be future-proof, as it may very well be called from a threaded
interrupt handler in the future.

This has the added benefit that it also tells GCC that the 'queue_len'
may be updated outside the visible scope, and therefore its value may
change at any moment. This fixes the code not working properly when
building with LTO.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>