Avoid conditionally-compiled code
Conditionally-compiled code (wrapped by #ifdef/#ifndef guards) is
problematic; because it is very difficult to compile-test all possible
combinations of the various flags verified by the guard macros.

Instead, evaluate compile-time constants, which will cause the compiler
to compile the exact same code independently of the configuration, which
means if it compiles for one configuration, it will compile for all
configurations. The compiler will then be smart enough to
garbage-collect dead code, dropping everything not included in the
current configuration.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>