examples: Allow building all examples in parallel
Instead of building all the examples sequentially in a single thread,
build all of them in parallel.

The Makefile will first generate a list of all the sub-folders that
contain a Makefile, and all of them will be added to the list of phony
rules.

The default target will then depend on that sub-directory list, running
Make recursively on each of those subdirectories, also using the
special $(MAKE) variable so that the same jobserver is used.

Because of that, all the Makefiles used only for recursion are useless
now and can be dropped.

On my PC (i7-10850H 12-core), when building with LTO running
'time make -j12 -C examples/dreamcast/' used to take about 8min 15s, and
now takes about 1min 24s, while compiling more examples, as previously
the gldc/ examples were not compiled.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>