Add support for gUSA atomic model
Add support for the gUSA atomic model, which is what is used in the
Linux kernel. By specifying -matomic-model=soft-gusa, the gUSA atomic
model can now be used.

Instead of creating critical sections by masking interrupts, gUSA relies
on the scheduler detecting atomic sections, identified by the stack
pointer going negative, and rolling back to the beginning of the atomic
section if interrupted in the middle.

The minimal overhead added to the scheduler is balanced by the gains
recovered by not masking interrupts for every single atomic access.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>