sq: Work around MMU issue by... disabling MMU
This change makes it possible to use SQs in a program that uses the MMU,
by disabling the MMU in sq_lock(), and re-enable it again in
mmu_restore().

This means that memory-mapped data must not be accessed while the SQ is
locked. This should be fine, as it wasn't working properly before, so
it's not a regression.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>