utils/pvrtex: fix and improve Makefile
We must not add KOS specific include paths when compiling a *host*
utility. This can cause the compilation to fail (as it did on my
system).

A lot of the rules are unnecesary, the default rules can be used as long
as the proper variable names are used.

The "all" rule is moved as the first rule, so that running "make" will
be equivalent to running "make all".

Signed-off-by: Paul Cercueil <paul@crapouillou.net>