sq: Update sq_lock() prototype to work with MMU
Change its prototype: it now returns a pointer, which is the translated
address of the source pointer. In non-MMU use, it's returns the same
value as SQ_MASK_DEST(src).

When MMU is enabled, sq_lock() will reset the address translation for
the SQ memory area, and convert the source address as an offset to that
memory area, making it possible to use SQs when MMU is enabled.

The caveat is that we only map one megabyte of SQ memory area for
simplicity, and therefore the SQ memory window is reduced.
This should be fine though, as apart from e.g. a whole-VRAM clear,
nobody will ever need such a big window.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>