pvr: Mark TA ready on PVR CORE render complete IRQ
Previously, the Tile Accelerator (TA) was deemed idle when the previous
frame was done rendering and flipped to the front buffer in a VBLANK
period.

This was fine for applications and games able to push more than 60 frames
per second, but it caused huge delays for those unable to reach this
threshold.

Address this issue by marking the TA as "ready" when all the lists have
been transferred and the PVR CORE is idle, and mark the PVR CORE as
"idle" after a frame has been moved to the front buffer.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>