kernel: Add performance monitor API
Add a new performance monitor API, built on top of the performance
counters API.

The performance monitor API is started with perf_monitor_init(), which
takes two performance events (perf_cntr_event_t) that will be monitored
aside from the call count and the total time, and is stopped with
perf_monitor_exit().

Then, a performance monitor can be added to any functional block using
the perf_monitor() macro. The performance will be monitored from the
moment the macro is called until the end of the functional block, and
the execution statistics of each call will be summed to total numbers.

Finally, perf_monitor_print() will display the collected information
about the performance monitors.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>