perf_monitor: Add macro for branch likeliness analysis
Introduce a new perf_monitor_if() macro.

This macro is designed to be used inside an "if" expression, for instance:

  if (perf_monitor_if(!strcmp("test", str))) {
     ...
  }

The resulting performance monitor will measure the number of calls, and
the number of times the branch was taken (in event1) and the number of
time it was not (in event0).

Signed-off-by: Paul Cercueil <paul@crapouillou.net>