pvr: Rework pvr_wait_ready()
The previous implementation used a semaphore, which was somewhat
problematic; it meant that there had to be a 1:1 ratio between the
number of calls to pvr_wait_ready(), and the number of renders by the
PVR. Having two successive calls, for instance, would cause the second
one to time out.

Rework the implementation so that pvr_wait_ready() will wait until the
PVR is ready if needed, but will always return straight away if the PVR
is free.

Instead of a semaphore, this new implementation uses the genwait
mechanism on the pvr_state.ta_busy object. This field has also changed
in its meaning, as it is now a generic "Tile Accelerator is busy" flag,
independently on whether or not the hybrid mode is in use.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>