pvr: Always wait for TA ready before trying to use it again
Attempting to use the Tile Accelerator for a new scene while it is busy
processing the previous scene will almost always result in a crash.

Call pvr_wait_ready() when starting a new scene to make sure that the TA
is idle.

Instead of calling it in pvr_scene_begin(), delay the call as much as
possible, to leave time for the TA to complete its task. This means
calling it in pvr_list_begin() for lists that use direct rendering, and
right before uploading the DMA data for lists that use DMA.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>