pvr: Rework and improve IRQ handling
The previous way of handling IRQs was weird at best. The PVR rendering
was not started directly after all lists were transferred by the
Tile Accelerator, but following the next VBLANK event.

This caused some games to be throttled down, as they had to wait for the
PVR rendering to be completed before starting a new scene, and the PVR
rendering was not started as soon as possible.

Address that by starting the PVR rendering as soon as all lists have
been transferred. In the case where the previous rendered scene hasn't
been flipped yet (which happens when the game tries to render faster
than the refresh rate), the PVR rendering is instead started from the
VBLANK interrupt handler.

Note that for simplicity the VBLANK handling code has been separated
from the main PVR IRQ handler.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>