fs_vmu: Improve handling of VMU file headers
- Add function fs_vmu_set_header().
  This function can be used to set a specific header (which contains the
  metadata, icons...) to an opened VMU file, replacing the one it previously
  had (if any).

- Add function fs_vmu_set_default_header().
  This function will set a default header, that will be used for new files
  which were not assignated their own header.

The header is now skipped when opening a VMU file (that has a header),
which is a breaking change from the previous behaviour of KallistiOS. It
is however kept internally, which means that a file opened read-write
will reuse its header (unless overwritten by fs_vmu_set_header()).

The point of this change is to hide the header from the application
completely, which can now read/write the VMU files with a proper header
as if they were PC files. It is then much easier for PC ports to
read/write savestates that look good.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>