fs_iso9660: Get rid of global FD table
Instead of using one big FD table with a fixed size, and using an index
into that table as the token passed among the backend functions, use a
pointer to the file handle structure as the token itself.

The FD table can then be removed, and replaced by a double-linked list
of the file handle structs, so that we can still keep track of what's
opened in case we need to invalidate all the handles at once.

This change removes the arbitrary limitation on the number of files that
can be opened with the iso9660 filesystem.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>