mutex: Move recursive/errcheck tests outside critical section
When the lock's holder is the current thread, we do not need IRQs to be
disabled to update the lock's counter, because a thread is not going to
race with itself.

Note how mutex_lock() now calls mutex_trylock() to factorize the code.
The (!m->count) check with IRQs disabled is then done twice, for now,
but this will change once mutex_trylock() uses an atomic
compare-and-swap to lock the mutex.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>