mutex: Wake up all threads waiting on a lock
In the case where multiple threads are waiting on a lock, waking up
only one thread does work, as the awaken thread will then get the lock,
do its business then wake up the next one in line.

The problem with that is that the genwait algorithm will always wake up
the thread that waited on the lock the longest, and not the one with the
highest execution priority.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>