CI: Add autobuilder for KallistiOS
The autobuilder will build all the toolchains we have a profile for,
*if* something in dc-chain or in the CI script changed.

In the case where the CI is triggered by a merge to the master branch
(aka. not a pull request), those Docker images will be pushed to
Docker Hub.

Those images will then be reused to build KallistiOS itself, with every
toolchain we have a profile for.

In the case where nothing changed in dc-chain, the total build time is
below 2 minutes. If the toolchains need to be built, the total build
time is below 50 minutes.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>