CI: Add autobuilder for KallistiOS
Build all the toolchains we have a profile for, as Docker images.
The toolchains are only built when toolchain-specific or CI files
change, which means that they are only rebuilt when needed.

In a second step, those toolchains are used to build KallistiOS itself.
When the toolchains don't need to be updated, only KallistiOS is built,
which is extremely fast (around 2 minutes vs. 40+ minutes for a full
build).

When a PR is merged to the master branch, and the toolchains need to be
updated, the Docker images will be pushed to Docker Hub, which is used
to host the latest toolchain images.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>