dc-chain: Rework Docker script
Rework the Docker script, so that it will use the current KOS tree
instead of downloading the master branch from the upstream repository.

This makes it possible to generate Docker images from different
repositories, different branches, or even with local changes.

Additionally, install the "bison" dependency which is needed to build
the most recent toolchains.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>