Fix built object file not added to the KOS library
The source files inside the kernel/ subdirectory (which is just
version.c for now) were compiled, but the object files were not added to
the KallistiOS library.

Address this by adding a Makefile dependency on the "defaultall" rule,
which will copy the object files to the build folder, so that they are
subsequently compiled inside the KallistiOS library.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>