pvr: Update pvr_get_front_buffer()
Previously, the VRAM offset of the front buffer was doubled, and then
masked to the VRAM size. This effectively meant that for a front buffer
at VRAM address 0x400000, we would get VRAM address 0x0. This actually
works, as the front buffer lies in 32-bit space while textures lie in
64-bit space; the framebuffer at that address could then be used as a
texture from VRAM address 0x0 with specifically crafted U/V values.

The problem with the previous implementation is that due to the masking,
the calling code has no way to know if the front buffer resides in the
lower 4 MiB of VRAM, or the upper 4 MiB of VRAM - and this information
is needed for using the front buffer as a texture.

Update the function so that it does clamp the front buffer address to
the VRAM size before the doubling, and not after. With the example
above, the address computed for the front buffer would be 0x800000,
which is bigger than the VRAM size. This does not matter, as the poly
compile functions will clamp the texture address to the VRAM size. But
it will allow the calling code to know if the front buffer is in the low
or high VRAM.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>