Add definition for tu32_t tuple type
Tuple types are useful especially as return values; this is because on
SH4 the ABI allows a function to return two 32-bit values, one in r0,
the other in r1. As a result, the functions that need to return two
values do not need to do so through return value pointers, which
generally cause the compiler to generate pretty bad code (due to
aliasing etc.).

Signed-off-by: Paul Cercueil <paul@crapouillou.net>