thread: Reschedule on genwait wakeup if needed
If the currently running thread has a lower priority than the thread(s)
we woke up, we now reschedule.

This makes sense especially if we consider that the current thread could
be the idle thread, in which case it does not make any sense to wait for
the next timer event to reschedule.

If the scheduler's timer is not running, we also need to reschedule to
make sure that the scheduler will start the timer, if we have more than
one thread running.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>