dc-chain: Fix command removing fake-kos.o requiring valid $PATH
The command removing fake-kos.o from the C library used sh-elf-gcc-ar,
expecting it to find it in the $PATH, as its parent directory is added
there inside variables.mk.

However, I am not sure that modifying PATH as a Makefile variable has
any effect on the actual $PATH environment variable. Case in point, the
command using sh-elf-gcc-ar to remove fake-kos.o from the C library
would only execute if the toolchain's path was already in the $PATH
variable of the build environment.

This caused all kinds of nasty issues.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>