dc-chain: Fix removal of fake-kos.o from C library
On some systems it appears that the Makefile's $(shell ) function won't
search the extra paths added to a PATH variable overriden inside the
Makefile, even if it is exported.

Address this issue by passing the absolute paths to the tools needed to
remove the fake-kos.o from the C library.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>