CI: Only run build-kos step for supported platforms
We only need to run the build-kos step for the platforms supported by
KallistiOS. Add a 'build-kos-check' step which then allows to
conditionally run the build-kos step.

The .github/build-kos.sh script was already checking for that, but doing
so before the build-kos step is executed means that we're only
downloading the Docker image of the toolchain if we will use it.

Also dropped the install of 'coreutils' in build-kos.sh as it's already
done in the Docker script.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>