timer: Use compile-time-known offset for TNS array
Instead of reading back the scaling factor (TPSC value) from the
hardware registers and use it as the offset into the TNS array, use the
compile-time-known TIMER_TPSC macro as the offset, because we know for
sure this is the same value.

The compiler will then be able to optimize the code better.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>