irq: Don't save/restore VBR field during context switches
The VBR field is not updated at any point by any thread (nor should it
be). Therefore, we can skip saving/restoring it when switching contexts.

This actually matters, as the IRQ context struct is 228 bytes, which is
4 bytes too big to fit in exactly 7 cache lines. By moving the VBR field
to the end, and ignoring it during context switches, we can improve
their speed.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>